Fix variable type to match printf format specifier in test_stack
* tests/test_stack.c (run_one_test): Change type of "index" local
variable from long to int to match printf format specifier (when
VERBOSE defined); cast from "arg" pointer to integer via size_t (to
avoid 64-bit compiler warning).

+++ tests/test_stack.c
@@ -137,7 +137,7 @@ volatile AO_t ops_performed = 0;
 void * run_one_test(void * arg)
 {
   list_element * t[MAX_NTHREADS + 1];
-  long index = (long)arg;
+  int index = (int)(size_t)arg;
   int i;
   int j = 0;
